Provide reasonable error message for non-OC refs
When getting a non-OC reference attribute value from the connector,
midPoint now rejects it with a more reasonable error message.

Furthermore, "attributes to get" computation algorithm was fixed so that
strategy=IMPLICIT no longer causes non-default-returned attribute to
be explicitly requested (in the presence of attribute with MINIMAL
fetch strategy).