Check for conflicts for single-valued item
As described in MID-9621, sometimes it is not enough to check
the conflicts (multivalue to single-valued items) using the computed
delta alone. The reason is that the existing value may not get into
the delta.

This commit adds a functionality to check this explicitly while
consolidating.

It can present a non-compatible change for some environments; but
definitely it's bugfix.