Fix previewChanges with associations inbounds
Preview changes should provide results analogous to the real execution.
This would involve simulating the execution of deltas computed
in individual projection waves. Historically, this was not done so;
only the simulations feature (which is something like advanced preview
changes) does this in a limited way.

We don't dare to add something like this to previewChanges, to avoid
breaking existing functionality.

But, for the associations to work, we need to have at least generated
IDs in assignments. This commit does exactly so.

It is a kind of ugly hack, but the whole previewChanges feature is
now like that. We should probably replace it with simulations in
the future.

This commit also adapts TestAssociationsFirstStepsActiveDirectory
to be closer to the real scenarios used in the training.