Check trigger existence in optimizing creator
The OptimizingTriggerCreator did not check for trigger existence
in a strict way, which lead to creation of duplicate triggers
on some occasions - e.g., in cluster or when caches were cleared
in the meanwhile.

Although not considered harmful initially, it seems that this
behavior can lead to relatively massive creation of triggers
in some environments.

Hence, this commit strengthens the trigger existence checks,
at the cost of some extra repository read accesses.

Related to MID-9368.