Improve support for existence=true setting
The "existence=true" setting in schema handling means that we want
to keep the shadow existing even if focal object is being deleted.

However, the default behavior of midPoint is that it considers all
focus data values to be removed, which - by default - means that
the projector is going to remove all values driven by outbound
mappings.

This may be desired behavior for some data (e.g., when anonymizing
shadows belonging to deleted users), but not for all.

This commit adds easily accessible "midpoint.isFocusDeleted()"
method that can be used in conditions for those mappings that control
attributes that have to be kept intact on user deletion.

Related to MID-9669.

(cherry picked from commit 9c285062bc21b294fbd902c4152b4032f0d48829)