Fix various caching-related issues
1. When doing clockwork operation with projection reconciliation, the
USE_CACHED_OR_FRESH option is now specially handled even at the context
loading time: The current status of the shadow is inspected, and if it's
not fresh enough, the full shadow is loaded (just like in midPoint 4.8).

2. Inbound mappings for password and activation no longer require the
full shadow to be loaded. The respective items' caching status is
checked instead.

3. Caching metadata are correctly put into shadow even if it's retrieved
from the resource (i.e., without noFetch option).

4. Tests in model-impl were adapted to the caching.