Relax parsing stale repo shadows
When resource schema changes in an incompatible way, schema issues can
occur during repo shadow reading. This commit tries to deal with that:

- On one hand, we want to work with the correct data. Even skipping
unreadable items can lead to a disaster, e.g., if there are mappings
that propagate this information further.

- On the other hand, we do not want to fail hard when the resource
schema changes.

The probably simplest solution is to be strict for fresh shadows,
but more relaxed for stale ones: the data will not be used anyway
(except for the identifiers - but they do not change as frequently).

The general recommendation will be: after you change the schema, please
invalidate the cache.

This fixes a couple of provisioning tests.