Clarify legacy associations vs object types
!!! In order to work properly, simulated associations should be known
at the level of object classes. This was true from the beginning,
but now, with shadow caching, it is required even more. !!!

For associations (reference attributes) simulated via capabilities,
this is ensured by design. But legacy associations are traditionally
defined at the level of object types, not object classes.

The connection to object class definition is usually done by using
"default=true" setting for the object type where the association is
defined. (Although this brings some issues if there are other object
types of the same OC, but without the association - see e.g. MID-9591
and MID-9910.)

Hence, this commit:

1. Adds TestDummyIntents to explore the behavior of legacy associations,
so that we would exactly know how much they can be supported e.g. with
the new shadow caching.

2. Fixes TestUniversity in story tests (it was failing with caching
enabled) by specifying the association at the level of object class:
this time explicitly, without using default=true. (Resource definition
parsing code had to be slightly adapted for this.)