Preliminary fix for MID-9932
Attempts to fix the primary cause of this issue - the corruption
of (shared) definitions by the resource wizard, which stems for not
freezing beans used by the definitions.

However, there are other parts of the new (associations-related)
definitions that are still not frozen appropriately. They should
be reviewed and fixed.