Treat refs in unclassified shadows gracefully
When legacy associations are present, and the association definition
is not visible in the object class, the following may occur:

When storing repo shadows with references, the shadow gets unreadable:
it contains the reference, e.g., because it was read from resource
with kind/intent known, but the created shadow has only object class
name, not the kind+intent (as the classification was not done yet).

This commit introduces special lax mode that ignores the references
when parsing such shadows. (With a little hack related to ignoring
empty reference attributes that should not be there; repo should take
care of that.)

This fixes TestIntents running with shadow caching enabled.