Do various fixes related to shadow caching
1. Improved handling of inbound mappings w.r.t. executed deltas.
The problem is that as the current input value we sometimes need
the real current state (when there's no apriori delta), but if there
is a delta, we need the state before the delta. This was implemented
by enhancing LensObjectDeltaOperation; also, some clarifications
regarding execution waves and preview mode were done there.

2. Fixed reloading of shadows after deltas being executed. It is needed
to provide "shadow current" state for inbounds, as described above.
Note: it would be better done via existing 'fresh' flag, but that's
not quite possible now, so a hacked 'reloadNeeded' solution was created.
See MID-9944.

3. Using provisioning instead of repository when loading shadows
in MissingShadowContextRefresher.

4. Better handling of "projection gone" situations for activation and
credentials processing: NotLoadedException instead of (sometimes
misleading) ObjectNotFoundException, and catching the exception
in credentials processing.

5. Data dependencies between projections (MID-8929, see "dataBinding"
config property) are resolved by explicitly loading the dependent
shadow from the resource. Maybe shadow invalidation would be better;
but this is to be seen.

6. Unrelated: Fixed OperationResult handling at some places.

This fixes some failing tests, in particular TestMultiResource (while
also adapting the tests slightly) and TestMappingInbound.