Fix "membership inbound sync disabled" behavior
When:

1. an assignment was created via inbounds (from an association value),
2. later the target shadow is turned to "outbound only",
3. the respective association value was removed from the user account;

then the resulting behavior was wrong: the respective assignment was
removed, because there was no longer the association value we could
ask about the synchronization policies!

This is now fixed: Any assignment that is to be removed because of
range is extra checked whether there is a compatible shadow (although
no longer connected to the user account), and checks that its inbound
membership sync is not disabled.