Fix inbounds regarding deltas
#1 in 441ac351a6b6c8c28be8baeafcfb5a85b2b80757 was perhaps too naive,
as it did not take volatile resources in to account: the old state
of the object plus delta may be different from its new state.

So, here we are more cautious: we assume that if there's no delta
for an attribute, the new state is the authoritative one. It's still
not 100% correct (as there may be deltas on volatile attributes), but
it's the best that could be done here.

This should fix TestVolatility and TestInternalCorrelationMultiAccounts.