Treat unmanaged group removal gracefully
When unmanaged group is removed, the role is gone. So the assignments
are dangling. After this commit, user recomputation (with reconciliation
option on) as well as group reconciliation should both remove these
extra assignments nicely.

Technically, we no longer fail on exceptions in
doesVetoTargetValueRemoval method in <associationSynchronization>
expression evaluator.

Should resolve MID-9917.