Migrate previewChanges to simulations, part 1
The previewChanges functionality was created long before simulations
were born, and it significantly overlaps them. Moreover, it's quite
problematic in nature, as it does not simulate the execution closely
enough. This leads to a lot of tickets, like MID-9853.

So this commit migrates previewChanges functionality to use simulations.
It is a drop-in replacement: Model API remains unchanged, only
the implementation is switched.

There are some areas where we depend on the original behavior, though;
at least for now. So, the previewChangesLegacy method was created.
It is currently used for requesting access and for computing direct
and indirect assignments. This is to be reviewed soon.

Also removed (experimental) simulation options in ModelExecuteOptions.

Work in progress. Maybe we should migrate all nuances from legacy
previewChanges method to the modern one.

---
Related to MID-9853.