Fix createTargetObjectsFilter methods
In principle, 2bc9eacc70eaf55a7db129df948c1c12ffcad88a was correct, but
the filters produced cannot be used for on-resource (or even noFetch)
searches.

So this commit provides the "resourceSafe" flag that provides
filters that are less precise, but at least executable.

Related to MID-10023. Should fix failing TestEntitlements, TestScience,
and TestUnix/TestUnixTolerantAux.