Adapt TestFirstSteps
As a side effect of 96b51b1aa696e96b6eddaaff15184de7a5b84c17,
inbounds are now executed for accounts being created (in simulation
mode) on a resource.

This may cause troubles if there are "magic" attributes, which are
computed by the resource. In TestFirstSteps, such an attribute is
ri:uid, which is used to provide a value for user name.

So an additional condition had to be added to this attribute inbound
mapping.

This makes TestFirstSteps work again.

Related to MID-10039 and MID-10051.