Extend mapping specification in value metadata
The association type name and the shadow tag were added to the mapping
specification in the provenance metadata. The former is primarily
to restrict the scope of the required uniqueness of mappings names.
The latter is to enable automatic ranges for multi-account scenarios.

This commit provides extensive tests in TestBasicValueMetadata,
including the one regarding auto-ranges in multivalued extension items
in assignments (import from associations).

Note that deleting shadows is not natively supported yet
(values and/or their metadata are not removed).

Also:
 - added "name" as the natural key for association type definition.