Ignore no-expression mappings for associations
Currently, inbound/outbound mappings without expressions in association
types make no sense: they are interpreted as "asIs" mappings, but fail
because of missing source (outbound case) or type incompatibility
(association->assignment, inbound case).

This troubles background tasks when resource configuration is edited
in the GUI, as the first step is to create+save such empty mappings, and
only later they are filled-in by an expression.

Hence, this commit makes such empty associations mappings to be ignored.

It is questionable whether it's the correct way, though. In the future,
when we'll provide a reasonable default behavior for such mappings,
we might want these empty mappings to be effective.

Related to MID-10076.