Relax task un-pausing for multinode activities
When dealing with coordinator-worker task relation (used to implement
multinode activities), we can allow the coordinator to be resumed under
more relaxed conditions than we can do in the general case.

The reason is that the coordinator is able to decide whether to continue
waiting or to close itself more accurately, by looking at the work
state.

This brings more resiliency into situations where (some of) worker tasks
cannot be successfully closed, for example, because of their node going
down.

This fixes MID-10021.