Fix inbounds from deleted shadows
The current behavior of inbound mappings related to shadows being
deleted or unlinked is currently like this:

1. When the shadow is explicitly deleted, inbounds mappings
are evaluated. The corresponding focus values are removed.

2. When the shadow is unassigned or unlinked, inbound mappings
are ignored. The corresponding focus values are kept.

(See newly-added TestInbounds.test400-430. It is maybe slightly
inconsistent, but this behavior is there for a long time.)

However, the current implementation broke down on resource
unavailability combined with shadow caching: the inbounds for case #2
above were executed right on the next clockwork click after
the deletion, resulting in focus values removal.
(The TestConsistencyMechanism.test610 uncovered this.)

This commit fixes the behavior by explicitly skipping inbounds for
contexts that were [attempted to be] deleted, even if the deletion
was not executed yet by the provisioning module.

Other changes: adapted parts of TestConsistencyMechanism (story)
to caching.

Makes TestConsistencyMechanism work with caching enabled.
Also, resolves MID-10109.