Remove "modern" shadow password caching
The caching of passwords in shadows is not usable yet. Only the hashing
method was available; and bringing encryption there needs resolving
some tough questions, like how should we select the applicable security
policy that would drive the storage method.

So, this commit removes that functionality, at least for now.
Legacy password caching can still be used, if needed.

Related to MID-10050.