Implement password caching
The password caching was not implemented as part of the shadow caching
feature provided in midPoint 4.9. It is implemented now. Please see
https://docs.evolveum.com/midpoint/devel/design/password-caching-4.9.1/
for more information.

Related behavior changes in this commit:

- The somewhat chaotic behavior of cached vs retrieved password value
(hashed vs empty/hashed/encrypted) requires more elaborate approach
to comparing password values in ObjectModificationConstraintEvaluator.

- MidPoint no longer provide INCOMPLETE flag for attributes passed
from the LDAP connector when "tolerantMultivalueReduction" feature
is turned on. Please see
https://docs.evolveum.com/midpoint/devel/design/incomplete-items-4.9.1/

Unrelated changes:

- Slightly updated credentials documentation.
- Minor code improvements at various places.
- Activation data is erased from the cache when an account is fetched
and the caching is disabled.
- Re-enabled some methods in TestOpenDj (forgotten in disabled state).
- Fixed checking whether password capability is enabled when asking
about "isPasswordReturnedByDefault".

Related to MID-10050 and MID-9958 (duplicate).