Use volatilityTrigger instead of disabling caching
This is the more correct way of dealing with hidden dependencies between
attributes (dn -> uid in this case).

(It is not the comprehensive solution, as there are attributes that
are autonomously generated on account creation, like internalId in the
dummy resource. We cannot handle these using volatilityTrigger setting.)

See MID-10051.