Performance: Use version from table column, not fullObject
Storing version in fullObject requires us to recompute fullObject
even on changes to linkRef or assignments, which are not stored
in full object anymore.

If the cannonical version is stored in column, this would allow us
to not fully parse full object in deltas not affecting items stored
in full object.