Improve shadow attribute volatility support
This commit adds new schema for attribute volatility, specifying it as
relations between attributes (source->target).

It provides a limited implementation, covering three cases:

- legacy volatility trigger: anything may change when attr X changes,
- "all" volatility trigger: anything may change when any attr changes,
- volatility on create: given attributes may be set/modified on create.

Related:

- Comprehensive support for volatility testing on the dummy resource:
ConnectorOperationHook objects.

- Clarified the ADD operation at the shadows and resource objects levels
so that created attributes and volatile ones are now written to BOTH
high-level and lower-level representations of the object to be added.

As part of this, the auditing is improved now: we audit the actual
object that was sent to the connector plus everything we got back
from the resource (in add operation + when fetching volatiles).

- Simplified return values from ADD, MODIFY, and DELETE operations at
both UCF and "resource objects" levels. The handling of these operations
at the "shadows" levels was simplified as well.

- A bug preventing native reference modifications in "avoid duplicate
values" mode was fixed.

- A rather obsolete code for invoking failure notifications for a narrow
set of schema-related errors in ADD operation was removed. It
complicated the code flow, and was rather an exception than a rule:
nothing like that exists for other operations. If needed, we should
implement the functionality in more generic way.

(cherry picked from commit b883f583f6a5ab35c1706845cd4887f02830792b)