Fix UCF->ConnId conversion for referenced objects
The original algorithm for converting midPoint (UCF) objects to ConnId
was too eager: it converted the whole structure of connected objects
in full. This was useless, and sometimes lead to failures - when
referenced objects contained unresolved references.

Now we convert only what is strictly required: root object and embedded
objects in full, but only the identifiers for referenced objects.

This resolves MID-10271.