MID-10231: Fixed deadlock in Sqale repository by processing URIs before opening transaction
In case of auto-partitioning URI caching transactions and retries did not worked
corretly which resulted in deadlock on PostgreSQL side.

URI caching is done for some URIs before real add transaction is opened.
This ensures URIs will be available to transaction and less retries.

The solution does not account for all possible cached URIs since some
are cached via Item Mappings and some are hardcoded in the code.