Fix cleanup after finish to remove also subtasks
**What**

Remove not just root task, but also all its subtasks when it is
cleanedup.

**Why**

When cleanup after finish is set on task, it should be auto removed
after specified duration from its completion.

However there was a bug, which caused, that only the root task was
removed. If task run spawned any persistent subtasks (e.g. when
number of workers per node is specified), those were not cleaned up.
This change fixes it.