Fix searching shadows in the repository
When a shadow was found in the repository, the definition applied to it
was taken from the query, not from the shadow itself. This could cause
problems e.g. with the association definitions.

This commit applies the correct definition, relevant for the shadow
in question.

(Somewhat) related to MID-10285.

(cherry picked from commit d3d8831cd71c3801924ade3431f7399ef65deaaa)