Treat dangling references in expressions better
When a reference was used in an expression in the default "object
variable mode", and its value was not resolvable, a fatal error
occurred. This was not correct according to the docs.

This commit fixes that. The null value is used instead. Note that
the behavior is still not quite consistent, as described in MID-10296.

This resolves MID-10162.