Remove getRepresentativeObjectDefinition method
This method was present on reference and association definitions.
It was dangerous, as it was not clear what was the "representative
object definition" - especially for more complex configurations.

The method was replaced by "getDefinitionForTargetObjectClass" method
that is much more clear. (There is still a bit of confusion, but that's
in midPoint for a long time; it's good enough for now. See MID-10309.)

Other changes:

- Shadow consistency checks were strengthened a bit.

Related to MID-9932.