Simplify volatility schema (temporarily)
In order to simplify the GUI, this commit reduces the cardinality
of inbound/outbound dependencies (in attribute volatility spec)
from 0..N to 0..1. The idea is that in current version, the content
of one item is just a single multi-valued property, so the whole
item can be safely single-valued.

As soon as the ShadowItemDependencyType get more content (like
source/target paths, as planned), the inbound/outbound items will
be returned to multi-valued state. It will be a compatible change then.

(cherry picked from commit 37bd03b66e8990086d0f777cd48fd7f194fa1947)