Optimize attribute definitions lookups
The findLocalItemDefinition method for resource object definitions
(and a couple of similar methods) are more efficient now: instead of
iterating through definitions, map lookups are used.

This improves the performance of the pre-processing in
provisioning-impl, especially for shadows with larger # of attributes.

Related to 415eaad8a2e9bf971bbc150a143e01372ad0ffca in prism.