Optimize attribute definitions lookups
The findLocalItemDefinition method for resource object definitions
(and a couple of similar methods) are more efficient now: instead of
iterating through definitions, map lookups are used.

This improves the performance of the pre-processing in
provisioning-impl, especially for shadows with larger # of attributes.

Related to ca024ec6365cafeb7edd0645d49cd9b165f3404c in prism.