Reduce the number of query cache invalidations
Invalidation of the global (and sometimes local) query cache may be
a costly operation, as we must walk through all the cached queries.
This commit ensures that it's skipped in cases when the object type
in question (like ShadowType) is not query-cached at all.

To be able to do that, this commit forbids caching ObjectType queries
(which are not recommended anyway).

This should improve the performance of shadow add, modify, and delete
operations.