Avoid extra SQL access for embedded shadows
Embedded shadows (used for associations and reference attributes)
do not need "operationExecution" information. So, we exclude it
from the respective repo search operations.

(The back side is that the repository cache is avoided. But we assume
that shadows are usually not cached anyway.)

This should improve performance of provisioning "get" and "search"
operations when many entitlements are present.