Remove max password length constraint
**What**

Remove constraint on maximum password length from default configuration.

**Why**

It basically goes against our documented best practice. The limitation
was there because of the defaults in some external systems (AD). However
we should not relent on the security just because of external systems
default policies, thus this change removes the constraint.

**Fixes**: MID-10305