Add "ignoreWhenGenerating" value policy feature
This commits adds the possibility of accepting selected characters in
a value policy, but never use them when generating the values.

It is used to stop problematic characters (space, '#', '&', '*',
apostrophe, and double quotes) being generated by the default value
policy.

This resolves MID-9541.

(Somewhat cherry-picked from d1057fe278a7efdc6414fa2decb15e545372caa0;
but the ValuePolicyProcessor code is totally different, so the change
had to be implemented from scratch.)