Make TestSystemPerformance less random
In order to compare performance between various runs more exactly,
this commit reduces the variability of generated data.

Previously, the number of operations varied mainly because the number
of technical roles (and therefore, entitlements) could be different
amount of overlap between randomly-generated and randomly-assigned
business roles.

Now we reduce the randomness by using a generator with a fixed seed.