Optimize reviewer notifications from certification
**What**

Optimize and simplify the queries (and code) used to select reviewers
who should receive "required review" notifications.

**Why**

The previous implementation tended to cause opening of certification
campaign stages to hang indefinitely\*. These freezes seem to be caused
by a large number of inserts followed by more complex select.

It looks like rapid inserts causes need for the DB to do vacuum/analyze
operations, but those operations are run periodically with certain
delay. If the select operation hits the DB before the "autovacuuming"
period (called nap time) elapses, it may appear to freeze\* in the DB
without any sign of real activity.

Lowering the "autovacuum nap time" configuration option seems to
mitigate this issue. By default, this option is set to 1 minute. During
local testing, setting it to 10 seconds helped (although this was only
limited manual testing).

\*We are not sure if this freeze is just temporary and the select would
eventually finish or if it gets stuck forever.

**Notes**

Originally, the `notifyReviewers` method executed two queries. The first
query was simple and did not cause any issues. However, the second
query, which involved joins over three different tables (or in fact,
nested selects in an `exists` clause), caused problems.

After several iterations, I decided to use a single query to load all
cases with open work items, map these to reviewers in code, and then
send notifications based on that mapping. Open work items are those
whose `closeTimestamp` is null, generally meaning they belong to the
currently opened stage.

This approach has pros and cons:

**Disadvantages:**
- Loads a large number of cases into memory
- Filtering happens in code (no fancy index optimizations)

**Advantages:**
- Only one DB round trip (though technically not entirely true due to
  iterative search)
- Much simpler query
- The same cases are not read more than once for reviewers with open
  work items

Note that when a notification is sent, all of the reviewer's cases from
the current stage are included, not just those where the reviewer has
work items without replies. This is one of the reasons I chose to load
all the cases.

Regarding the iterative search, in this case it probably has only one
benefit and that is it does not have limit on number of returned items
in total. The non iterative search has limit of 10 000 records.

**Fixes**: MID-10225