Try to reduce the number of cloning operations
1. We now use READ_ONLY option to retrieve objects at more places.
The most prominent one is the focus objects that enter the clockwork.
(There's no point in fetching shadows as READ_ONLY because they are
always modified in the provisioning module.)

2. This commit also fixes a couple of places in the projector/clockwork
where the current object was unnecessarily modified.

A prominent exception is the injection of value metadata into data
coming from resources. This is important for midPrivacy Data provenance
prototype. But it had to be temporarily disabled, see MID-10514.

3. Fixed handling of LensElementContext-related authorizations in
DataAccessProcessor: treats also immutable objects.

4. Other minor optimizations were done, like avoiding re-reading
a shadow when its synchronization situation is going checked whether
it's up to date (in most typical cases). We simply rely on the already
loaded data there.

5. Removed objectBefore from ModifyObjectResult (repo API). It was not
needed, and it required an extra clone on each repo modification op.

6. !!! Changed contract of ModificationsSupplier in dynamic object
modifications (repo API) !!! The original approach was that the caller,
i.e., repository, cloned the existing object before handling it to the
callback (modifications supplier). Now, the client is responsible for
not modifying the provided object in any way.

Work in progress. Some tests may fail.