Improve performance of computing "complete" flag
This flag is used to know whether we can return an object (obtained
using standard options), when asking for a complete object.

The computation of this flag was not efficient enough. It is now fixed:
Materialization of value metadata is no longer done, and the double
computation (for local and global cache) is avoided.