Fix OpenDJ related conntests
We can no longer assert on the synchronization situation being
DELETED in the dead shadow returned from the model.getObject
operation. It was a side-effect of the "discovery" operation
on that dead shadow. Recent code changes related to performance
optimizations cause that the discovery operation is carried out
on a copy of the shadow, hence none of its effects propagate
to the shadow being returned.

The shadow returned is (still) marked as dead, anyway, and this is
the key information for the client.

The synchronization situation is still stored in the repository.