Fix preview changes for hidden projections
Projections with denied access no longer cause the preview changes
operation to fail: If a user has no authorization to see particular
projection (shadow), the "preview changes" operation used to finish
with "Access denied" fatal error even if there were parts of the result
visible to the user. This is now fixed: only the relevant projections
are hidden now. All the remaining data are displayed to the user.

This fixes also the same behavior for request access functionality,
where the "preview changes" is invoked as part of conflicts preview.

Resolves MID-10397.