Implement formatter for object deltas
**What**

Implement formatting of added/deleted/modified objects (and containers)
which is based on the "Visualisation Abstraction".

**Why**

Currently we have two kinds of deltas "visualizations".

One is in form of a preview in the UI, which is based on top of the
"Visualization Abstraction", that provides human readable descriptions
of certain changes.

The second form is plain text, used in notifications. That is based
directly on deltas and does not contain many features which the
"Visualization Abstraction" provides (e.g. human readable descriptions
akka overviews).

This formatters are implemented as an effort to unify all deltas
visualizations to be based on the same abstraction and thus to provide
more or less the same set of visualization "features". They are meant
for usage in notifications, but they may also be used in scripts via
existing `TextFormatter`.

**Notes**

This is still work in progress and implementation may/will change (see
al the TODOs and FIXMEs).

Just one note aside of the TODOs/FIXMEs is the new `IDeltaFormatter`
interface. The `I` prefix will be removed. It is there for now just to
have this commit compilable without the need of changing a lot of other
files (`DeltaFormatter` already exists and it is the current
implementation based on deltas).

**Implements**: MID-6112