Add hiding of specified path in delta visualizations
**What**

- Add option to hide specified item paths from delta visualization.
- Few other changes in visualizations, which helps with future code
  changes (e.g. do not return wild-card types when possible).

**Why**

Formatters used by notifications (and possibly also from scripts) had
this option before. In order to be backward compatible would need such
behavior also in new visualization based formatters.

But I believe, that better place to implement such hiding is directly in
the Visualization abstraction, so it could be potentially used also by
other code, not just formatters.

Regarding of the removal of wild-card return types. In general it is not
recommended to return wild-card types if possible, because it in lot of
cases forces clients of the API to cast returned objects.

**Notes**

Currently there is one limitation regarding hiding of the parameters
from deltas. If the delta contains only changes which should be hidden,
then it can cause that visualization would be returned as if something was
changed, but it will not contain any parameters. That is not very
desired and it could potentially fall under MID-5350. This issue was
present also in old formatters (mitigated to some extend). There
already is an improvement for this - MID-6111.