Relax freezing objects when R/O option is used
Recently, we've introduced a behavior that always froze objects being
returned from "get" and "search" repository operations if read-only
option is used.

This may be safer, but generally it is a waste of processing time.

This commit returns to the original semantics of the read-only option:
the repository cache may return either mutable or immutable version
of the object. (The latter case occurs if the object is being put or
is retrieved from the cache.)

Some extra "freeze" calls were added at places where the returned
objects are to be shared.